[add] save if track entry was prediction, and set max successive pred in processing
now it is possible to distinguish a track entry created based on a kalman prediction
from one created from an actual measurement. the ncc will be negative if the entry
corresponds to a kalman prediction.

in the processing tab it is possible to set the max number of successive kalman
predicted track entries that are accepted. if the robot is lost (and it's position
predicted by kalman) for over X frames, then the track entries from the moment the
robot was lost to the moment it was found again are discarded for coverage analysis.
setting this limit to 0 will ignore all entries corresponding to kalman predictions.